refactor project structure: move template models to new folder
