--- internal/models/gomod.go
@@ -1,18 +1,0 @@
-package models
-
-const (
-	GOMOD_TEMPLATE = "go.mod.tmpl"
-	GOMOD_FILE     = "go.mod"
-)
-
-type GoModInfo struct {
-	GoVersion   string
-	PackageName string
-}
-
-func NewGoModInfo(packageName string) *GoModInfo {
-	return &GoModInfo{
-		PackageName: packageName,
-		GoVersion:   GoVersion(),
-	}
-}
